+++ nekoyume/Assets/_Scripts/Lib9c/Model/Quest/GoldQuest.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
+using System.Numerics;
 using Bencodex.Types;
 using Nekoyume.Model.EnumType;
 using Nekoyume.TableData;
@@ -13,7 +14,7 @@ namespace Nekoyume.Model.Quest
     {
         public readonly TradeType Type;
 
-        public GoldQuest(GoldQuestSheet.Row data, QuestReward reward) 
+        public GoldQuest(GoldQuestSheet.Row data, QuestReward reward)
             : base(data, reward)
         {
             Type = data.Type;
@@ -45,11 +46,13 @@ namespace Nekoyume.Model.Quest
 
         protected override string TypeId => "GoldQuest";
 
-        public void Update(decimal gold)
+        public void Update(BigInteger gold)
         {
             if (Complete)
                 return;
 
+            // FIXME: _current를 BigInteger로 바꾸는 게 좋지 않을까요…
+            // 이대로라면 overflow로 돈을 2^32 NCG 이상 벌면 같은 퀘스트 두 번 이상 깰 수 있을 듯.
             _current += (int)gold;
             Check();
         }
